// Code/Websites/DanpheEMR/wwwroot/DanpheApp/src/app/reporting/shared/depertment-wise-stat-report.dto.ts
export class DepartmentwiseStatReport_DTO {
    DepartmentName: string = null;
    FollowupFemaleAdult: number = 0;
    FollowupFemaleChild: number = 0;
    FollowupMaleAdult: number = 0;
    FollowupMaleChild: number = 0;
    NewFemaleAdult: number = 0;
    NewFemaleChild: number = 0;
    NewMaleAdult: number = 0;
    NewMaleChild: number = 0;
    FollowupOthersChild: number = 0;
    FollowupOthersAdult: number = 0;
    NewOthersChild: number = 0;
    NewOthersAdult: number = 0;
    Total: number = 0;
}